w3c-reports/12092014/mac/runned-tests/selectors-api/tests/submissions/Opera/level2-baseline-expected.txt

CONSOLE MESSAGE: line 22: TypeError: null is not an object (evaluating 'parent.appendChild')
This test requires JavaScript.

Harness Error. harness_status.status = 1 , harness_status.message = TypeError: null is not an object (evaluating 'parent.appendChild')
